Moved kiosk stuff into own folder
