+++ custom_components/fusion_solar/fusion_solar/kiosk/kiosk.py
