+++ custom_components/fusion_solar/fusion_solar/kiosk/kiosk_api.py
@@ -3,7 +3,7 @@ import logging
 import html
 import json
 
-from .const import (
+from ..const import (
     ATTR_DATA,
     ATTR_FAIL_CODE,
     ATTR_SUCCESS,
