+++ custom_components/fusion_solar/sensor.py
@@ -11,8 +11,8 @@ from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from .fusion_solar.const import ATTR_DATA_REALKPI, ATTR_REALTIME_POWER, ATTR_TOTAL_CURRENT_DAY_ENERGY, \
     ATTR_TOTAL_CURRENT_MONTH_ENERGY, ATTR_TOTAL_CURRENT_YEAR_ENERGY, ATTR_TOTAL_LIFETIME_ENERGY
-from .fusion_solar.kiosk import Kiosk
-from .fusion_solar.kiosk_api import FusionSolarKioksApi, FusionSolarKioskApiError
+from .fusion_solar.kiosk.kiosk import Kiosk
+from .fusion_solar.kiosk.kiosk_api import FusionSolarKioksApi
 from .fusion_solar.energy_sensor import FusionSolarEnergySensorTotalCurrentDay, \
     FusionSolarEnergySensorTotalCurrentMonth, FusionSolarEnergySensorTotalCurrentYear, \
     FusionSolarEnergySensorTotalLifetime
